pyproject.toml
[project]
name = "md-note-taking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Django>=4.2",
    "djangorestframework",
    "python-markdown",
    "language-tool-python",
    "markdown>=3.10",
]
